+++ src/FSharp.Management.PowerShell/HostedRuntime.fs
@@ -3,47 +3,93 @@
 open System
 open System.Management.Automation
 open System.Management.Automation.Runspaces
+open System.Threading
+open System.Security.Principal
+
+open TypeInference
 
 type IPSRuntime =
-    abstract member AllCmdlets  : unit -> TypeInference.PSCommandLet[]
+    abstract member AllCmdlets  : unit -> PSCommandLet[]
     abstract member Execute     : string * obj seq -> obj
     abstract member GetXmlDoc   : string -> string
 
 /// PowerShell runtime built-in current process
-type PSRuntimeHosted(snapIns:string[]) =
-    let runSpace = 
-        let config = RunspaceConfiguration.Create(); 
-        for snapIn in snapIns do
-            if (not <| String.IsNullOrEmpty(snapIn)) then 
-                let info, ex = config.AddPSSnapIn(snapIn)
-                if ex <> null then
-                    failwithf "AddPSSnapInException: %s" ex.Message
-        let rs = RunspaceFactory.CreateRunspace(config)
-        rs.Open()
-        rs
-    let cmdlets =
-        runSpace.RunspaceConfiguration.Cmdlets
-        |> Seq.map (fun cmdlet ->
-                        let wrapper = TypeInference.createPSCommandLet cmdlet
-                        wrapper.RawName, wrapper)
-        |> Map.ofSeq
+type PSRuntimeHosted(snapIns:string[], modules:string[]) =
+//  let modules = [|"Azure"; "ActiveDirectory"|]
+//  let snapIns = [|""|]
+    let runSpace =
+        try
+            let initState = InitialSessionState.CreateDefault()
+
+            for snapIn in snapIns do
+                if not <| String.IsNullOrEmpty(snapIn) then
+                    let _, ex = initState.ImportPSSnapIn(snapIn)
+                    if ex <> null then
+                        failwithf "ImportPSSnapInExceptions: %s" ex.Message
+
+            let modules = modules |> Array.filter (String.IsNullOrWhiteSpace >> not)
+            if not <| Array.isEmpty modules then
+                initState.ImportPSModule(modules);
+
+            let rs = RunspaceFactory.CreateRunspace(initState)
+            Thread.CurrentPrincipal <- GenericPrincipal(GenericIdentity("PowerShellTypeProvider"), null)
+            rs.Open()
+            rs
+        with
+        | e -> failwithf "Could not create PowerShell Runspace: '%s'" e.Message
+
+    let commandInfos =
+        let ps  = PowerShell.Create(Runspace=runSpace)
+        //Get-Command -CommandType @("cmdlet","function") -ListImported
+        ps.AddCommand("Get-Command")
+          .AddParameter("CommandType", ["cmdlet"; "function"])
+          .AddParameter("ListImported")
+          .Invoke()
+        |> Seq.toArray
+
+    let commands =
+        try
+            commandInfos
+            |> Seq.map (fun x ->
+                match x.BaseObject with
+                | :? CommandInfo as ci -> ci
+                | w -> failwithf "Unsupported type of command: %A" w)
+            |> Seq.map (fun cmd ->
+                match cmd with
+                | :? CmdletInfo | :? FunctionInfo->
+                    if cmd.ParameterSets.Count > 0 then
+                        seq {
+                            for pSet in cmd.ParameterSets do
+                                let wrapper = createPSCommandLet cmd pSet
+                                yield wrapper.UniqueID, wrapper
+                        }
+                    else
+                        failwith "Cmdlet/Function is not loaded: %A" cmd
+                | _ -> failwithf "Unexpected type of command: %A" cmd
+            )
+            |> Seq.concat
+            |> Map.ofSeq
+        with
+        | e -> failwithf "Could not load cmdlets: %s\n%s" e.Message e.StackTrace
+
     interface IPSRuntime with
-        member this.AllCmdlets() =
-            cmdlets |> Map.toSeq |> Seq.map snd |> Seq.toArray
-        member this.Execute(rawName,parameters:obj seq) =
+        member __.AllCmdlets() =
+            commands |> Map.toSeq |> Seq.map snd |> Seq.toArray
+        member __.Execute(uniqueId, parameters:obj seq) =
             // Create command
-            let cmdlet = cmdlets.[rawName]
-            let command = Command(cmdlet.RawName)
+            let cmdlet = commands.[uniqueId]
+            //TODO: Execute commands using PowerShell class
+            let command = Command(cmdlet.Name)
             parameters |> Seq.iteri (fun i value->
-                let key,ty = cmdlet.ParametersInfo.[i]
+                let key, _,ty = cmdlet.ParametersInfo.[i]
                 match ty with
                 | _ when ty = typeof<System.Management.Automation.SwitchParameter>->
-                    if (unbox<bool> value) then 
+                    if (unbox<bool> value) then
                         command.Parameters.Add(CommandParameter(key))
                 | _ when ty.IsValueType ->
                     if (value <> System.Activator.CreateInstance(ty))
                     then command.Parameters.Add(CommandParameter(key, value))
-                | _ -> 
+                | _ ->
                     if (value <> null)
                     then command.Parameters.Add(CommandParameter(key, value))
             )
@@ -54,22 +100,30 @@ type PSRuntimeHosted(snapIns:string[]) =
 
             // Format result
             let tys = cmdlet.ResultObjectTypes
-            let tyOfObj = TypeInference.getTypeOfObjects tys result
+            let tyOfObjOpt = TypeInference.getTypeOfObjects tys result
 
-            let collectionConverter = 
-                typedefof<TypeInference.CollectionConverter<_>>.MakeGenericType(tyOfObj)
-            let objectCollection = 
-                if (tyOfObj = typeof<PSObject>) then box result 
-                else result |> Seq.map (fun x->x.BaseObject) |> box
-            let typedCollection = 
-                collectionConverter.GetMethod("Convert").Invoke(null, [|objectCollection|])
+            match tyOfObjOpt with
+            | None -> box None
+            | Some(tyOfObj) ->
+                let collectionConverter =
+                    typedefof<TypeInference.CollectionConverter<_>>.MakeGenericType(tyOfObj)
+                let objectCollection =
+                    if (tyOfObj = typeof<PSObject>) then box result
+                    else result |> Seq.map (fun x->x.BaseObject) |> box
+                let typedCollection =
+                    collectionConverter.GetMethod("Convert").Invoke(null, [|objectCollection|])
 
-            if (tys.Length = 1) 
-            then typedCollection
-            else let ind = tys |> Array.findIndex (fun x-> x = tyOfObj)
-                 let funcName = sprintf "NewChoice%dOf%d" (ind+1) (tys.Length)
-                 cmdlet.ResultType.GetMethod(funcName).Invoke(null, [|typedCollection|])
-        member this.GetXmlDoc(rawName:string) =
+                let choise =
+                    if (tys.Length = 1)
+                    then typedCollection
+                    else let ind = tys |> Array.findIndex (fun x-> x = tyOfObj)
+                         let funcName = sprintf "NewChoice%dOf%d" (ind+1) (tys.Length)
+                         cmdlet.ResultType.GenericTypeArguments.[0]
+                            .GetMethod(funcName).Invoke(null, [|typedCollection|])
+
+                cmdlet.ResultType.GetMethod("Some").Invoke(null, [|choise|])
+
+        member __.GetXmlDoc(rawName:string) =
             // Create command
             let command = Command("Get-Help")
             command.Parameters.Add(CommandParameter("Name", rawName))
@@ -83,10 +137,10 @@ type PSRuntimeHosted(snapIns:string[]) =
                 prop.Value
             match result with
             | [|help|] ->
-                let lines = 
+                let lines =
                     let description = (help?description :?> obj [])
-                    if description = null then [||] 
-                    else 
+                    if description = null then [||]
+                    else
                         description
                         |> TypeInference.CollectionConverter<PSObject>.Convert
                         |> List.toArray