+++ src/FSharp.Management.PowerShell/Runtime.fs
@@ -5,16 +5,16 @@ open HostedRuntime
 open ExternalRuntime
 
 // PowerShell runtime resolver
-let private runtimes = ref Map.empty<string[] * bool, IPSRuntime>
-let Current(snapIns, is64bitRequired, isDesignTime) =
-    let key = (snapIns, isDesignTime)
+let private runtimes = ref Map.empty<string[] * string[] * bool, IPSRuntime>
+let Current(snapIns, modules, is64bitRequired, isDesignTime) =
+    let key = (snapIns, modules, isDesignTime)
     if (not <| Map.containsKey key !runtimes) then
-        let value = 
+        let value =
             if (is64bitRequired && not(Environment.Is64BitProcess)) then
-                if (isDesignTime) 
-                then new PSRuntimeExternal(snapIns) :> IPSRuntime
+                if (isDesignTime)
+                then new PSRuntimeExternal(snapIns, modules) :> IPSRuntime
                 else failwith "You should compile your code as x64 application"
-            else 
-                PSRuntimeHosted(snapIns) :> IPSRuntime
+            else
+                PSRuntimeHosted(snapIns, modules) :> IPSRuntime
         runtimes := (!runtimes |> Map.add key value)
     (!runtimes).[key]