+++ src/FSharp.Management/AssemblyInfo.fs
@@ -4,9 +4,9 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("FSharp.Management")>]
 [<assembly: AssemblyProductAttribute("FSharp.Management")>]
 [<assembly: AssemblyDescriptionAttribute("Various type providers for management of the machine.")>]
-[<assembly: AssemblyVersionAttribute("0.1.1")>]
-[<assembly: AssemblyFileVersionAttribute("0.1.1")>]
+[<assembly: AssemblyVersionAttribute("0.2.1")>]
+[<assembly: AssemblyFileVersionAttribute("0.2.1")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "0.1.1"
+    let [<Literal>] Version = "0.2.1"