+++ tests/FSharp.Management.Tests/PowerShellProvider.Tests.fs
@@ -4,25 +4,52 @@ open FSharp.Management
 open NUnit.Framework
 open FsUnit
 
-type PS = PowerShellProvider< PSSnapIns="", Is64BitRequired=false >
+let [<Literal>]PSSnapIns = "" 
+let [<Literal>]Modules = "Microsoft.PowerShell.Management;Microsoft.PowerShell.Core"
 
-[<Test>] 
-let ``Get system drive``() = 
-    let items = PS.``Get-Item`` [|@"C:\"|] 
-    items |> should haveLength 1
-    let dir = items.Head.BaseObject :?> System.IO.DirectoryInfo
-    dir.FullName |> shouldEqual @"C:\"
+type PS = PowerShellProvider< PSSnapIns=PSSnapIns, Modules=Modules, Is64BitRequired=false >
+
+[<Test>]
+let ``Get system drive``() =
+    match PS.``Get-Item``(path=[|@"C:\"|]) with
+    | Some(Choice5Of5 dirs) ->
+        dirs |> should haveLength 1
+        (Seq.head dirs).FullName |> shouldEqual @"C:\"
+    | _ -> failwith "Unexpected result"
 
 [<Test>]
 let ``Check that process `testtest` does not started``() =
-    match PS.``Get-Process``(name = [|"testtest"|]) with
-    | Choice4Of4(processes) -> 
-        processes |> should be Empty
+    match PS.``Get-Process``(name=[|"testtest"|]) with
+    | None -> ignore()
     | _ -> failwith "Unexpected result"
-    
+
 [<Test>]
-let ``Get list of registered snapins``() = 
+let ``Get list of registered snapins``() =
     match PS.``Get-PSSnapin``(registered = true) with
-    | Choice1Of2(snapins) -> 
+    | Some(snapins) ->
+        snapins.IsEmpty |> should be False
+    | _ -> failwith "Unexpected result"
+
+
+type PS64 = PowerShellProvider< PSSnapIns=PSSnapIns, Modules=Modules, Is64BitRequired=true >
+
+[<Test>]
+let ``Get system drive x64``() =
+    match PS64.``Get-Item``(path=[|@"C:\"|]) with
+    | Some(Choice5Of5 dirs) ->
+        dirs |> should haveLength 1
+        (Seq.head dirs).FullName |> shouldEqual @"C:\"
+    | _ -> failwith "Unexpected result"
+
+[<Test>]
+let ``Check that process `testtest` does not started x64``() =
+    match PS64.``Get-Process``(name=[|"testtest"|]) with
+    | None -> ignore()
+    | _ -> failwith "Unexpected result"
+
+[<Test>]
+let ``Get list of registered snapins x64``() =
+    match PS64.``Get-PSSnapin``(registered = true) with
+    | Some(snapins) ->
         snapins.IsEmpty |> should be False
     | _ -> failwith "Unexpected result"